(refactor) TEST 패키지 분리
