+++ src/test/java/domain/generator/RandomNumberGeneratorTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.generator;
 
+import domain.generator.RandomNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
