+++ src/test/java/domain/hint/StrikeTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.hint;
 
+import domain.hint.Strike;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
