+++ src/test/java/domain/number/SourceNumbersTest.java
@@ -1,6 +1,7 @@
-package domain;
+package domain.number;
 
-import domain.SourceNumbers;
+import domain.number.BaseBallNumber;
+import domain.number.SourceNumbers;
 import exception.BaseBallException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
